Add files via upload
The script split.py was used to divide the gene sequence from homo.fa (it couldn't be uploaded as it contains gene sequence of length 1 billion+ and exceeds the file size limit of GitHub) into 168 files, each containing sequences of equal length (4242), for the negative directory. Maintaining a uniform gene sequence length across all files is necessary to implement the sliding window approach.